Remove SigningPool and other cruft
