+++ examples/flows/flows.ts
@@ -19,7 +19,6 @@ import {
   uploadFile,
 } from '@monaxlabs/aspen-sdk/dist/apis';
 import { ContractService, Currency } from '@monaxlabs/aspen-sdk/dist/apis/publishing';
-import { asyncYield, ISSUER_ROLE, SigningPool } from '@monaxlabs/aspen-sdk/dist/apis/utils/signing-pool';
 import { ClaimBalance, getClaimBalances } from '@monaxlabs/aspen-sdk/dist/claimgraph';
 import {
   Address,
@@ -35,7 +34,7 @@ import {
 import { parse } from '@monaxlabs/aspen-sdk/dist/utils';
 import { BigNumber, BigNumberish, Signer } from 'ethers';
 import { providers } from 'ethers/lib/ethers';
-import { formatEther, parseEther } from 'ethers/lib/utils';
+import { formatEther } from 'ethers/lib/utils';
 import { createReadStream } from 'fs';
 import { GraphQLClient } from 'graphql-request';
 import { URL } from 'url';
@@ -205,36 +204,12 @@ async function cmdDirectIssueB(): Promise<void> {
   const accounts = generateAccounts(maxSize, { mnemonic: demoMnemonic, provider });
 
   const gasStrategy = await getGasStrategy(provider);
-  const pool = await SigningPool.fromAccessControlContract(
-    signer.privateKey,
-    contract.address,
-    {
-      maxSize: maxSize,
-      provider,
-      seed: '0x303af788d4200da2c2918709ee3b6f871a108f0b42aaebb8b761077aa2c634c6',
-      replenish: {
-        reserve: signer,
-        topUp: parseEther('0.01'),
-        lowWater: parseEther('0.001'),
-        gasStrategy,
-      },
-      logger: console.error,
-    },
-    ISSUER_ROLE,
-  );
-
-  await pool.isInitialised();
 
   const promises: Promise<NFTTokenIssuance[]>[] = [];
   for (const account of accounts) {
-    await asyncYield();
-    promises.push(
-      pool.do(async (signer) => {
-        const signerAddress = await signer.getAddress();
-        console.error(`Calling issue to receiver ${account.address} with signer ${signerAddress}`);
-        return contract.issueNFT.must(signer, { to: account.address, quantity: 1 }, gasStrategy);
-      }),
-    );
+    const signerAddress = await signer.getAddress();
+    console.error(`Calling issue to receiver ${account.address} with signer ${signerAddress}`);
+    await contract.issueNFT.must(signer, { to: account.address, quantity: 1 }, gasStrategy);
   }
   const issues = (await Promise.all(promises)).flat();
