--- examples/stripe/README.md
@@ -1,51 +1,0 @@
-# Stripe Webhook & Aspen Token Issuance
-
-This example receives a Stripe webhook on a completed payment from a `payment link`. This example assumes you have a
-working Stripe tenancy already available and that `accept terms` is either disabled, or handled prior to payment
-processing.
-
-We use the payment link metadata to define the collection & token to be issued. The Client Reference ID passed to the
-payment link should be the wallet which will receive the token.
-
-## Overview
-
-The webhook reciver is a basic Express Typescript app, using the Stripe and Aspen SDKs. The receiver:
-
-- accepts webhooks from Stripe,
-- checks the call signature,
-- performs basic payment validation, and
-- issues a single token via the Aspen Publishing interface.
-
-## Collection Setup
-
-Create a new 1155 collection using Aspen CreatorTools. Make a note of the collection guid. If crypto payments are not
-required ensure the proper claim setup.
-
-## Stripe Configuration
-
-1. Create a new `product` for each token (https://dashboard.stripe.com/test/products)
-2. For each product create a new `payment link` (https://dashboard.stripe.com/test/payment-links). Set `Quantity` to 1 and disable `Adjustable Quantity`. Each payment link may have a different payment amount. Under the `payment link` Metadata specify the following:
-
-- `collectionGuid` : The collection guid from CreatorTools.
-- `tokenId` : The token ID to be issued upon payment.
-
-3. Make a copy of the URL for each payment link.
-
-## Testing the webhook receiver
-
-For testing purposes we can self-host the webhook receiver and use the `stripe cli` to act as a webhook proxy.
-
-1. Install the Stripe CLI following the instructions at https://stripe.com/docs/stripe-cli.
-2. Launch the webhook proxy with `stripe listen --forward-to http://localhost:4344/webhook`.
-   - If you have changed the port number update as appropriate.
-3. Populate `credentials.json` and set the following environment variables:
-   - `STRIPE_APIKEY` - the secret key from https://dashboard.stripe.com/test/apikeys
-   - `STRIPE_ENDPOINTSECRET` - the `webhook signing secret` output in step 2.
-   - `SERVEDEMO` - `true` to enable the demonstration frontend
-   - `PORT` (Optional) - the port the webhook receiver will operate on.
-4. `cd demosite && pnpm install && pnpm build` (Optional if demo site is required)
-5. `pnpm start`
-
-## Productionisation
-
-In a production environment the webserver should be run on a publically accessible webserver and a permanant webhook should be configured in the Stripe Developers portal. The `STRIPE_ENDPOINTSECRET` environment variable should be set to the `Signing secret` generated by Stripe.
