--- examples/stripe/package.json
@@ -1,22 +1,0 @@
-{
-  "type": "module",
-  "scripts": {
-    "build": "tsc --build",
-    "tsnode": "ts-node --esm --experimental-specifier-resolution=node",
-    "start": "pnpm run tsnode server.ts",
-    "stripe:proxy": "docker run --rm -it stripe/stripe-cli:latest listen RjJqlmwZe007EfvlNrN --forward-to http://localhost:4344/webhook"
-  },
-  "dependencies": {
-    "@monaxlabs/aspen-sdk": "workspace:*",
-    "express": "^4.18.2",
-    "stripe": "^11.5.0"
-  },
-  "devDependencies": {
-    "@types/express": "^4.17.15",
-    "@types/node": "^18.11.18",
-    "concurrently": "^7.6.0",
-    "nodemon": "^2.0.20",
-    "typescript": "4.7.4",
-    "ts-node": "^10.9.1"
-  }
-}
