+++ sdk/src/apis/utils/index.ts
@@ -6,5 +6,4 @@ export * from './files';
 export * from './gating';
 export * from './issue';
 export * from './providers';
-export * from './queue';
 export * from './waiter';
